move dark mode context up
