+++ src/layouts/app.tsx
@@ -1,13 +1,11 @@
-import React, { useEffect, useState, createContext, useContext } from "react";
+import React, { useState } from "react";
 import { Outlet, redirect, useLoaderData } from "react-router-dom";
 import { fireproof } from "use-fireproof";
 import { authResult } from "../auth";
 import { SYNC_DB_NAME } from "../pages/databases/show";
-import Sidebar from "../components/Sidebar"
-import Header from "../components/Header"
-import { Button } from "../components/Button"
-
-const DarkModeContext = createContext()
+import Sidebar from "../components/Sidebar";
+import Header from "../components/Header";
+import { Button } from "../components/Button";
 
 const reservedDbNames: string[] = [
   `fp.${SYNC_DB_NAME}`,
@@ -58,18 +56,6 @@ async function getIndexedDBNamesWithQueries(): Promise<
   }
 }
 
-export function useDarkMode() {
-  const context = useContext(DarkModeContext)
-  if (!context) throw new Error('Darkmode context used outside of the Provider')
-  return context.isDarkMode
-}
-
-function getDarkMode() {
-  const storedDarkMode = localStorage.getItem("darkMode");
-  const darkModePreference = window.matchMedia("(prefers-color-scheme: dark)").matches
-  return storedDarkMode === "true" || darkModePreference;
-}
-
 export default function Layout() {
   const { databases, user } = useLoaderData<{
     databases: { name: string; queries: any[] }[];
@@ -78,7 +64,6 @@ export default function Layout() {
 
   const [showEmailModal, setShowEmailModal] = useState(user.publicMetadata.marketingOptIn === undefined);
   const [emailPreference, setEmailPreference] = useState(false);
-  const [isDarkMode, setIsDarkMode] = useState(() => getDarkMode());
 
   const handleEmailPreference = async () => {
     try {
@@ -101,45 +86,43 @@ export default function Layout() {
   };
 
   return (
-    <DarkModeContext.Provider value={{ isDarkMode }}>
-      <div className="flex h-screen w-full overflow-hidden bg-fp-bg-00">
-        {showEmailModal && (
-          <div className="fixed inset-0 backdrop-blur-sm z-50 flex items-center justify-center text-fp-p">
-            <div className="bg-fp-bg-00 border border-fp-dec-00 px-8 py-6 rounded-fp-s shadow-lg max-w-lg w-full">
-              <h2 className="text-34 mb-4">Email Preferences</h2>
-              <p className="mb-4 text-fp-s text-16">Would you like to receive emails from us?</p>
-              <div className="flex items-start gap-2 mb-6">
-                <input
-                  type="checkbox"
-                  id="emailPreference"
-                  checked={emailPreference}
-                  onChange={(e) => setEmailPreference(e.target.checked)}
-                  className="w-[28px] h-[28px] cursor-pointer mt-[-2px] accent-fp-a-02"
-                />
-                <label htmlFor="emailPreference" className="text-16 text-fp-s cursor-pointer hover:text-fp-p">
-                  Yes, I'd like to receive (occasional, genuinely informative) emails from Fireproof.
-                </label>
-              </div>
-              <div className="flex justify-end gap-2">
-                <Button
-                  variation="primary"
-                  type="button"
-                  onClick={handleEmailPreference}
-                >
-                  Confirm
-                </Button>
-              </div>
+    <div className="flex h-screen w-full overflow-hidden bg-fp-bg-00">
+      {showEmailModal && (
+        <div className="fixed inset-0 backdrop-blur-sm z-50 flex items-center justify-center text-fp-p">
+          <div className="bg-fp-bg-00 border border-fp-dec-00 px-8 py-6 rounded-fp-s shadow-lg max-w-lg w-full">
+            <h2 className="text-34 mb-4">Email Preferences</h2>
+            <p className="mb-4 text-fp-s text-16">Would you like to receive emails from us?</p>
+            <div className="flex items-start gap-2 mb-6">
+              <input
+                type="checkbox"
+                id="emailPreference"
+                checked={emailPreference}
+                onChange={(e) => setEmailPreference(e.target.checked)}
+                className="w-[28px] h-[28px] cursor-pointer mt-[-2px] accent-fp-a-02"
+              />
+              <label htmlFor="emailPreference" className="text-16 text-fp-s cursor-pointer hover:text-fp-p">
+                Yes, I'd like to receive (occasional, genuinely informative) emails from Fireproof.
+              </label>
+            </div>
+            <div className="flex justify-end gap-2">
+              <Button
+                variation="primary"
+                type="button"
+                onClick={handleEmailPreference}
+              >
+                Confirm
+              </Button>
             </div>
           </div>
-        )}
-        <Sidebar databases={databases} />
-        <div className="flex flex-col flex-1 overflow-hidden">
-          <Header user={user} isDarkMode={isDarkMode} setIsDarkMode={setIsDarkMode} />
-          <main className="flex-1 overflow-y-auto p-main">
-            <Outlet context={{ user }} />
-          </main>
         </div>
+      )}
+      <Sidebar databases={databases} />
+      <div className="flex flex-col flex-1 overflow-hidden">
+        <Header user={user} />
+        <main className="flex-1 overflow-y-auto p-main">
+          <Outlet context={{ user }} />
+        </main>
       </div>
-    </DarkModeContext.Provider>
+    </div>
   );
 }
